Cb edit user pony

+++ ponyproject/ponyapp/views/user_pony/user_pony_details.py
@@ -10,7 +10,10 @@ def get_pony(user_pony_id):
 
 # @login_required
 def user_pony_details(request, user_pony_id):
+    print("This is NOT a GET")
     if request.method == 'GET':
+        print("This is a GET")
+
         user_pony = get_pony(user_pony_id)
 
         template = 'user_ponies/detail.html'
@@ -20,14 +23,33 @@ def user_pony_details(request, user_pony_id):
 
         return render(request, template, context)
 
-    if request.method == 'POST':
+    elif request.method == 'POST':
         form_data = request.POST
+        # Check if this POST is for editing a pony
+        if (
+            "actual_method" in form_data
+            and form_data["actual_method"] == "PUT"
+        ):
+            print("this is a print")
+            # # retrieve it first:
+            user_pony_to_update = UserPony.objects.get(pk=user_pony_id)
 
-    if (
-        "actual_method" in form_data
-        and form_data["actual_method"] == "DELETE"
-    ):
-        user_pony = UserPony.objects.get(pk=user_pony_id)
-        user_pony.delete()
+            # # Reassign a property's value
+            # user_pony_to_update.pony_id = form_data['pony_id']
+            user_pony_to_update.price = form_data['price']
+            user_pony_to_update.details = form_data['details']
+            user_pony_to_update.condition_id = form_data['condition']
 
-        return redirect(reverse('ponyapp:user_pony_list'))
+            # # Save the change to the db
+            user_pony_to_update.save()
+
+            return redirect(reverse('ponyapp:user_pony_list'))
+
+        if (
+            "actual_method" in form_data
+            and form_data["actual_method"] == "DELETE"
+        ):
+            user_pony = UserPony.objects.get(pk=user_pony_id)
+            user_pony.delete()
+
+            return redirect(reverse('ponyapp:user_pony_list'))
